[refactor] 이메일 인증 시 예외 추가
- 이미 인증이 완료된 이메일이라면 Exception 발생
