+++ src/main/java/org/board/exercise_board/User/Security/SecurityConfig.java
@@ -1,7 +1,8 @@
-package org.board.exercise_board.User.config;
+package org.board.exercise_board.User.Security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
@@ -10,6 +11,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
+@EnableMethodSecurity
 @EnableWebSecurity
 public class SecurityConfig {
 
@@ -26,15 +28,19 @@ public class SecurityConfig {
             AbstractHttpConfigurer::disable
         )
 
-//        .authorizeHttpRequests(
-//            (authorizeRequests) ->
-//                authorizeRequests
-//                    .requestMatchers("/","/**").permitAll()
-//                    .anyRequest().authenticated()
-//        );
+        //홈,로그인,회원가입 페이지는 로그인 없이 접근 가능
+        .authorizeHttpRequests(
+            (request) -> request
+                    .requestMatchers("/","/user/signup","/user/login").permitAll()
+                    .anyRequest().authenticated()
+        )
 
-        // 기본 로그인페이지 없애기
+
         .formLogin(
+//            (form) -> form
+//                .loginPage("/user/login")
+//                .defaultSuccessUrl("/", true)
+//                .permitAll()
             AbstractHttpConfigurer::disable
         );
