+++ src/main/java/org/board/exercise_board/User/exception/ErrorCode.java
@@ -14,6 +14,7 @@ public enum ErrorCode {
 
   // 인증 관련
   NOT_FOUND_USER(HttpStatus.BAD_REQUEST, "해당 사용자를 찾을 수 없습니다."),
+  ALREADY_VERIFIED(HttpStatus.BAD_REQUEST, "이미 인증이 완료되었습니다."),
   EXPIRATION_TIME_IS_OVER(HttpStatus.BAD_REQUEST, "인증 유효시간이 지났습니다.");
 
   private final HttpStatus httpStatus;
